Fix deployment that was broken in 3e8af78
What
===
Move the app.yaml into the main package.

Why
===
In commit 3e8af783aa249b25684771956b4ac74630a645a6 I moved the main
package into a subpackage. I didn't move the app.yaml and that resulted
in the App Engine runtime looking for the main package in the wrong
place and failing to startup the service. According to the App Engine
docs the app.yaml should be in the main package.

Learnings
===
I tested the commit that broke the deployment immediately after the
deploy but I guess I must have still been hitting the old version
because it worked fine. Learning being to test against the version
specific routes from now on. Because of the way the application was
failing this wasn't reported to the Error Reporting feature that App
Engine / Google Cloud supports and so I wasn't notified of the issue.